Upgrade to spree v4.1 and ros-apartment
