+++ .gitignore
@@ -4,6 +4,7 @@
 .DS_Store
 .idea
 .project
+coverage
 tmp
 nbproject
 *.swp
