+++ .travis.yml
@@ -1,11 +1,21 @@
+os: linux
+dist: bionic
+
 language: ruby
+
 rvm:
-  - 2.1
-  - 2.2
-sudo: false
-cache: bundler
-before_script:
-  - sh -e /etc/init.d/xvfb start
-  - export DISPLAY=:99.0
+  - 2.5
+  - 2.6
+
+gemfile:
+  - gemfiles/spree_4_0.gemfile
+  - gemfiles/spree_4_1.gemfile
+  - gemfiles/spree_master.gemfile
+
+jobs:
+  allow_failures:
+    - gemfile: gemfiles/spree_master.gemfile
+
+script:
   - bundle exec rake test_app
-script: bundle exec rspec spec
+  - bundle exec rake spec
