+++ Appraisals
@@ -1,0 +1,14 @@
+appraise 'spree-4-0' do
+  gem 'spree', '~> 4.0.0'
+  gem 'rails-controller-testing'
+end
+
+appraise 'spree-4-1' do
+  gem 'spree', '~> 4.1.0'
+  gem 'rails-controller-testing'
+end
+
+appraise 'spree-master' do
+  gem 'spree', github: 'spree/spree', branch: 'master'
+  gem 'rails-controller-testing'
+end
