+++ Versionfile
@@ -1,2 +1,1 @@
-"3.0.x" => { :branch => "master" }
-"2.4.x" => { :branch => "2-4-stable" }
+"4.1.x" => { :branch => "4-1-stable" }
