+++ gemfiles/spree_4_1.gemfile
@@ -1,0 +1,8 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "spree", "~> 4.1.0"
+gem "rails-controller-testing"
+
+gemspec path: "../"
