+++ gemfiles/spree_master.gemfile
@@ -1,0 +1,8 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "spree", github: "spree/spree", branch: "master"
+gem "rails-controller-testing"
+
+gemspec path: "../"
