+++ lib/spree_shared/version.rb
@@ -9,8 +9,8 @@ module SpreeShared
 
   module VERSION
     MAJOR = 0
-    MINOR = 9
-    TINY  = 1
+    MINOR = 10
+    TINY  = 0
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join('.')
