+++ lib/tasks/db.rake
@@ -15,14 +15,8 @@ namespace :spree_shared do
       puts "Loading seeds & sample data into database: #{db_name}"
       initializer.load_seeds
       initializer.load_spree_sample_data
-      # Need to manually create admin as it's not created by default in production mode
-      if Rails.env.production?
-        initializer.create_admin
-      end
-      # load some extra sample data for spree_fancy
-      initializer.load_spree_fancy_sample_data
-
-      initializer.fix_sample_payments
+      
+      initializer.create_admin
 
       puts "Bootstrap completed successfully"
     end
